requirements.txt
openai>=1.0.0
pyyaml>=6.0
beautifulsoup4>=4.12.0
